requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx[http2]
pydantic[email]
python-multipart
psycopg2-binary
python-dotenv
pyjwt # JWT for authentication Doc: https://fastapi.tiangolo.com/tutorial/security/oauth2-jwt/#about-jwt Teste: https://jwt.io/
passlib[bcrypt] # Password hashing Doc: https://fastapi.tiangolo.com/tutorial/security/oauth2-jwt/#about-password-hashing
bcrypt==4.0.1